+++ zoo-modules/shared-module/test.setup.spec.js
@@ -1,0 +1,10 @@
+const Nightmare = require('nightmare');
+
+before(() => {
+	global.nightmare = new Nightmare();
+	global.nightmare.goto('http://localhost:5000');
+});
+
+after(() => {
+	global.nightmare.halt();
+});